test: use page fixtures, reduce regex match on swap
